+++ src/lib/views/models/download_view_model.dart
@@ -1,13 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:mediadrip/services/youtube_downloader_service.dart';
+import 'package:mediadrip/views/models/view_model.dart';
 
-class DownloadViewModel extends ChangeNotifier {
+class DownloadViewModel extends ViewModel {
   final YoutubeDownloaderService _downloaderService = YoutubeDownloaderService();
+
   final TextEditingController inputController = TextEditingController();
   final TextEditingController outputController = TextEditingController();
 
   bool get isDownloadServiceRunning => _downloaderService.state != DownloaderState.idle;
   String get downloadButtonLabel => isDownloadServiceRunning ? 'Wait...' : 'Download';
+
+  DownloadViewModel({@required BuildContext context}) : super(context: context);
 
   void download([String address]) {
     _updateOutput('Attempting to download drip...', reset: true);